requirements.txt
Flask==2.1.0
plotly==5.3.1
pillow
PyQt5
pyopengl
numpy
